int -> unsigned
These should obviously never be negative, so let's switch the type.

+++ src/renderer/rendertarget.h
@@ -5,7 +5,9 @@
 
 class RenderTargetBase {
 protected:
-	RenderTargetBase(VkFormat format, VkImageType imageType, VkImageViewType imageViewType, int width, int height, int depth, int arrayLayers, int mipLevels, VkImageUsageFlags usage, VkImageAspectFlags aspect, VmaMemoryUsage memoryUsage, VmaAllocationCreateFlags allocFlags) :
+	RenderTargetBase(VkFormat format, VkImageType imageType, VkImageViewType imageViewType,
+	                 unsigned width, unsigned height, unsigned depth, unsigned arrayLayers, unsigned mipLevels,
+	                 VkImageUsageFlags usage, VkImageAspectFlags aspect, VmaMemoryUsage memoryUsage, VmaAllocationCreateFlags allocFlags) :
 		format(format),
 		width(width),
 		height(height),
@@ -17,7 +19,7 @@ protected:
 		imageCreateInfo.flags = 0;
 		imageCreateInfo.imageType = imageType;
 		imageCreateInfo.format = format;
-		imageCreateInfo.extent = { (uint32_t)width, (uint32_t)height, (uint32_t)depth };
+		imageCreateInfo.extent = { width, height, depth };
 		imageCreateInfo.mipLevels = mipLevels;
 		imageCreateInfo.arrayLayers = arrayLayers;
 		imageCreateInfo.samples = VK_SAMPLE_COUNT_1_BIT;
@@ -46,11 +48,11 @@ protected:
 public:
 	VkFormat getFormat() { return format; }
 
-	int getWidth() const { return width; }
-	int getHeight() const { return height; }
-	int getDepth() const { return depth; }
+	unsigned getWidth() const { return width; }
+	unsigned getHeight() const { return height; }
+	unsigned getDepth() const { return depth; }
 
-	int getArrayLayers() const { return arrayLayers; }
+	unsigned getArrayLayers() const { return arrayLayers; }
 
 	VkImage getImage() { return image; }
 	VkImageView getImageView() { return imageView; }
@@ -58,8 +60,8 @@ public:
 protected:
 	VkFormat format;
 
-	int width, height, depth;
-	int arrayLayers;
+	unsigned width, height, depth;
+	unsigned arrayLayers;
 
 	VkImage image;
 	VkImageView imageView;
@@ -68,7 +70,7 @@ protected:
 
 class ColorRenderTarget : public RenderTargetBase {
 public:
-	ColorRenderTarget(VkFormat format, int width, int height, int mipLevels = 1, VkImageUsageFlags usage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT) :
+	ColorRenderTarget(VkFormat format, unsigned width, unsigned height, unsigned mipLevels = 1, VkImageUsageFlags usage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT) :
 		RenderTargetBase(format, VK_IMAGE_TYPE_2D, VK_IMAGE_VIEW_TYPE_2D, width, height, 1, 1, mipLevels, usage, VK_IMAGE_ASPECT_COLOR_BIT, VMA_MEMORY_USAGE_AUTO, 0)
 	{
 	}
@@ -76,7 +78,7 @@ public:
 
 class DepthRenderTarget : public RenderTargetBase {
 public:
-	DepthRenderTarget(VkFormat format, int width, int height, VkImageUsageFlags usage = VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT) :
+	DepthRenderTarget(VkFormat format, unsigned width, unsigned height, VkImageUsageFlags usage = VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT) :
 		RenderTargetBase(format, VK_IMAGE_TYPE_2D, VK_IMAGE_VIEW_TYPE_2D, width, height, 1, 1, 1, usage, VK_IMAGE_ASPECT_DEPTH_BIT, VMA_MEMORY_USAGE_AUTO, 0)
 	{
 	}
@@ -84,11 +86,11 @@ public:
 
 class Texture2DArrayRenderTarget : public RenderTargetBase {
 public:
-	Texture2DArrayRenderTarget(VkFormat format, int width, int height, int arrayLayers, VkImageUsageFlags usage = VK_IMAGE_USAGE_SAMPLED_BIT | VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT) :
+	Texture2DArrayRenderTarget(VkFormat format, unsigned width, unsigned height, unsigned arrayLayers, VkImageUsageFlags usage = VK_IMAGE_USAGE_SAMPLED_BIT | VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT) :
 		RenderTargetBase(format, VK_IMAGE_TYPE_2D, VK_IMAGE_VIEW_TYPE_2D_ARRAY, width, height, 1, arrayLayers, 1, usage, VK_IMAGE_ASPECT_COLOR_BIT, VMA_MEMORY_USAGE_AUTO, 0)
 	{
 		arrayImageViews.reserve(arrayLayers);
-		for (int i = 0; i < arrayLayers; ++i) {
+		for (unsigned i = 0; i < arrayLayers; ++i) {
 			VkImageSubresourceRange subresourceRange;
 			subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
 			subresourceRange.baseMipLevel = 0;
